Split up bloating files
