+++ fbsimctl/FBSimulatorControlKit/Sources/Writer.swift
@@ -15,41 +15,3 @@ import Foundation
 public protocol Writer {
   func write(string: String)
 }
-
-/**
- Enum for defining the result of a translation.
- */
-public enum ActionResult {
-  case Success
-  case Failure(String)
-
-  func append(second: ActionResult) -> ActionResult {
-    switch (self, second) {
-    case (.Success, .Success):
-      return .Success
-    case (.Success, .Failure(let secondString)):
-      return .Failure(secondString)
-    case (.Failure(let firstString), .Success):
-      return .Failure(firstString)
-    case (.Failure(let firstString), .Failure(let secondString)):
-      return .Failure("\(firstString)\n\(secondString)")
-    }
-  }
-}
-
-extension ActionResult : CustomStringConvertible, CustomDebugStringConvertible {
-  public var description: String {
-    get {
-      switch self {
-      case .Success: return "Success"
-      case .Failure(let string): return "Failure '\(string)'"
-      }
-    }
-  }
-
-  public var debugDescription: String {
-    get {
-      return self.description
-    }
-  }
-}
